Update README.md
In my opinion you will need to have a more detailed introductory paragraph here- I have written an example couple of sentences which you can take or leave- this is to set the scene and explicate the vision of the text. 
I would also recommend either at the start or the end of the book a list of all the contributors from the podcast who are featured or helped edit/review the episodes - to give the people what they want!