Update 02-upper-extremity.Rmd
Sorry for delay this section was a bit of a struggle bus. 
Unclear for these chapters when we discuss the anatomy  ( same with carotid and I imagine subsequent chapters) whether we could embed a link to images or similar eg Teach Me Anatomy similar as I see Rutherford's is used as the basis and is referenced here. I don't  think that actually putting images in is necessarily the way forward but it would be good to actively look at an image when considering this anatomically dense area.